Assignment/requirements.txt
requests
bs4
matplotlib.pyplot